DL-LTI-Openshift/DL-LTI-Openshift/requirements.txt
future >= 0.15.2
python-ilorest-library==v4.5.0
pyVmomi
pyVim
urlparse2
jsonpatch
jsonpath_rw
ansible-vault
jsonpath-ng
requests
